Adding a new workflow # Changes made - Added a home page to choose one of two possible workflows - Added a new page for the virtual library (currently just a template library) - Added routing to navigate to different pages # Pending changes - Incorporate back-end with front-end - CSS finetuning
