+++ UI/src/App.css
@@ -1,5 +1,7 @@
 .App {
   text-align: center;
+  width: 100vh;
+  height: 100vh;
 }
 
 .App-logo {
